Fix jenkins version pinning issue
The version of jenkins was not being pinned correctly and still grabs the LTS version, which during testing also matched the expected version causing tests to errornously pass.

This change fixes the issue and ensures the correct pinned version of jenkins gets installed. It also tags the docker images created with the correct pinned version.

Solo: @smford